Add improved_model based on progressive_model
Add "build_stage_preparation" to correct the calculation of the
loss of input. loss_input should exclude the texture loss of the
last layer (level), because the input is not computed from
the gradient of the last layer (level). While the loss of
the last layer (level) contributes to the calculation of the
gradient input in the current stage. The loss of input and the
current gradient input share a part of the computation graph.

Add two arguments "grad-ksize" and "pad-type" to progressive_model.
"grad-ksize" is the size of conv kernels used to transform the
gradient input. "pad-type" is the padding mode for conv layers.

Rename "build_upsampling_nn" to "build_upsampling_nnconv" for
clarity. That is, the upsampling module is ended with a conv layer.
Change the upsampling method to "tf.tile".